rationalize sections/review listeners/meta cache at country route level
