+++ src/client/pages/Country/SectionWrapper/SectionWrapper.tsx
@@ -1,78 +1,22 @@
-import React, { useEffect, useLayoutEffect } from 'react'
-import { useParams } from 'react-router-dom'
+import React, { PropsWithChildren, useLayoutEffect } from 'react'
 
-import { AssessmentName } from 'meta/assessment'
-import { Sockets } from 'meta/socket'
-
-import { useAppDispatch } from 'client/store'
 import { useGetTableSections, useSections, useTableSections } from 'client/store/metadata'
-import { useOriginalDataPoint } from 'client/store/ui/originalDataPoint'
-import { ReviewActions } from 'client/store/ui/review'
-import { useUser } from 'client/store/user'
-import { useCountryIso } from 'client/hooks'
+import { useSectionRouteParams } from 'client/hooks/useRouteParams'
 import MessageCenter from 'client/components/MessageCenter'
-import { SocketClient } from 'client/service/socket'
 import { DOMs } from 'client/utils/dom'
 
-type Props = {
-  children: JSX.Element
-}
-
-type SectionParams = {
-  assessmentName: AssessmentName
-  cycleName: string
-  sectionName: string
-}
-
-const SectionWrapper: React.FC<Props> = (props) => {
+const SectionWrapper: React.FC<PropsWithChildren> = (props) => {
   const { children } = props
 
-  const { assessmentName, cycleName, sectionName } = useParams<SectionParams>()
-  const dispatch = useAppDispatch()
-  const countryIso = useCountryIso()
-  const user = useUser()
+  const { sectionName } = useSectionRouteParams()
   const sections = useSections()
   const tableSections = useTableSections({ sectionName })
-  const originalDataPoint = useOriginalDataPoint()
   useGetTableSections()
 
   useLayoutEffect(() => {
     // scroll to top
     DOMs.scrollTo()
   }, [sectionName])
-
-  // subscribe to section review status update
-  useEffect(() => {
-    const requestReviewStatusEvent = Sockets.getRequestReviewStatusEvent({
-      countryIso,
-      assessmentName,
-      cycleName,
-      sectionName,
-    })
-
-    const updateReviewStatus = () => {
-      dispatch(
-        ReviewActions.getReviewStatus({
-          countryIso,
-          assessmentName,
-          cycleName,
-          sectionName,
-          odpId: originalDataPoint?.id,
-        })
-      )
-    }
-
-    if (user) {
-      updateReviewStatus()
-      SocketClient.on(requestReviewStatusEvent, updateReviewStatus)
-    }
-
-    return () => {
-      if (user) {
-        SocketClient.off(requestReviewStatusEvent, updateReviewStatus)
-      }
-    }
-  }, [countryIso, assessmentName, cycleName, sectionName, user, dispatch, originalDataPoint?.id])
 
   if (!sections || !tableSections) return null
