+++ src/client/store/ui/review/actions/getReviewSummary.ts
@@ -2,11 +2,11 @@ import { createAsyncThunk } from '@reduxjs/toolkit'
 import axios from 'axios'
 
 import { ApiEndPoint } from 'meta/api/endpoint'
-import { CountryIso } from 'meta/area'
+import { AreaCode } from 'meta/area'
 import { ReviewSummary } from 'meta/assessment'
 
 type Params = {
-  countryIso: CountryIso
+  countryIso: AreaCode
   assessmentName: string
   cycleName: string
 }
