+++ src/meta/socket/sockets.ts
@@ -1,9 +1,9 @@
-import { CountryIso } from 'meta/area'
+import { AreaCode, CountryIso } from 'meta/area'
 import { Assessment, AssessmentName, Cycle } from 'meta/assessment'
 import { MessageTopic } from 'meta/messageCenter'
 
 const getRequestReviewSummaryEvent = (props: {
-  countryIso: CountryIso
+  countryIso: AreaCode
   assessmentName: AssessmentName
   cycleName: string
 }) => {
